requirements.txt
asgiref==3.8.1
Django==5.2
django-rest-auth==0.9.5
djangorestframework==3.16.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
